Refactored large component into modular parts for improved structure
